feat: implement varnishlogreceiver loop
Adds the VarnishlogReceiver struct that is responsible
to execute varnishlog-json, parse the output and turn
it into traces. In this patch, the basic loop is implemented:

- Start varnishlog-json
- Read the stdout
- Parse it via serde_json

varnishlog-json is executed in "request" grouping mode
so we receive all transaction for a single request in a
JSON array per line to prevent buffering or other correlation
methods. By default, backend and ESI requests are included.

The main method now starts the varnishlogreceiver worker.