feat: add standard attributes to spans (#9)
* feat: add update_span method for injecting attributes

the new method update_span allows to update an existing
span with custom attributes from VarnishTx instances.

The implementation currently moves the varnishotel_semconv
attributes to the new function and adds the error status (if any).

* feat: add events from timeline to span

adds events from the request timeline to the span. each event
is timed according to the float timestamp so timings can be
seen accurately.

* feat: add client & backend attributes

adds the NETWORK_PEER_* attributes to identify the calling
client and backend (ip and port).

* feat: add backend attributes

this change extends the semconv specification with two new attributes:

- varnish.backend.name
- varnish.backend.conn_reused

The are extracted from the "backend" type in requests. Currently, the
backend names are extracted as is without any normalization.

* feat: add standard attributes per span

Adds a range of standard HTTP attributes to every span to enrich
every single request with useful "access log" like attributes.

* feat: add span name for backends

the backend requests are now updated and normalized. By default, the
pattern is "Varnish to <backend> fetch".

For regular backends (defined in VCL verbatim), the name is taken as-is.
For VMODs that define their own backend structures, such as
vmod-dynamic, special regexes are applied.

* fix: update rust formatting